+++ lib/ember_cli/route_helpers.rb
@@ -3,12 +3,6 @@ require "ember_cli/html_constraint"
 module ActionDispatch
   module Routing
     class Mapper
-      def mount_ember_assets(app_name, to: "/")
-        dist_directory = EmberCli[app_name].paths.dist
-
-        mount Rack::File.new(dist_directory.to_s) => to
-      end
-
       def mount_ember_app(app_name, to:, **options)
         routing_options = options.deep_merge(
           defaults: { ember_app: app_name },
@@ -25,7 +19,9 @@ module ActionDispatch
             get("#{to}(*rest)", routing_options)
           end
 
-          mount_ember_assets app_name, to: to
+          dist_directory = EmberCli[app_name].paths.dist
+
+          mount Rack::File.new(dist_directory.to_s) => to
         end
       end
     end